Move QuestionListTest from parser to command - rename to StartTest
